refactor(graduate): 스케줄 관련 로직 entity로 일부 분리
