+++ apps/graduate/src/admin/entities/admin-schedule/model/index.ts
@@ -1,3 +1,2 @@
 export { useScheduleList } from './useScheduleList';
 export { useScheduleDetail } from './useScheduleDetail';
-export { useScheduleContent } from './useScheduleContent';
