+++ apps/graduate/src/admin/entities/admin-schedule/model/useScheduleDetail.ts
