--- apps/graduate/src/admin/pages/schedule/hooks/index.ts
@@ -1,4 +1,0 @@
-export { useCreateSchedule } from './useCreateSchedule';
-export { useUpdateSchedule } from './useUpdateSchedule';
-export { useDeleteSchedule } from './useDeleteSchedule';
-export { useUpdateScheduleContent } from './useUpdateScheduleContent';
