+++ apps/graduate/src/admin/pages/schedule/model/index.ts
@@ -1,1 +1,6 @@
 export type * from './schedule';
+export { useCreateSchedule } from './useCreateSchedule';
+export { useUpdateSchedule } from './useUpdateSchedule';
+export { useDeleteSchedule } from './useDeleteSchedule';
+export { useUpdateScheduleContent } from './useUpdateScheduleContent';
+export { useScheduleContent } from './useScheduleContent';
