+++ apps/graduate/src/admin/pages/schedule/model/useCreateSchedule.ts
