+++ apps/graduate/src/admin/pages/schedule/model/useDeleteSchedule.ts
