+++ apps/graduate/src/admin/pages/schedule/model/useScheduleContent.ts
