+++ apps/graduate/src/admin/pages/schedule/model/useUpdateScheduleContent.ts
