+++ apps/graduate/src/admin/shared/api/index.ts
@@ -1,4 +1,3 @@
-export * from './fetchAdminUsers';
 export * from './schedule';
 export * from './notice';
 export * from './file';
