+++ apps/graduate/src/admin/widgets/StudentAddModal/api/index.ts
@@ -1,2 +1,3 @@
 export * from './submitGraduationUser';
 export * from './submitGraduationUsersBatch';
+export * from './fetchAdminUsers';
