+++ apps/graduate/src/shared/hooks/index.ts
@@ -1,4 +1,3 @@
-export { default as useTableState } from './useTableState';
-export * from './schedule';
-export * from './notice';
 export * from './useToast';
+export { useNoticeDetail } from './useNoticeDetail';
+export { useNoticeList } from './useNoticeList';
