--- apps/graduate/src/shared/hooks/notice/index.ts
@@ -1,2 +1,0 @@
-export { useNoticeDetail } from './useNoticeDetail';
-export { useNoticeList } from './useNoticeList';
