--- apps/graduate/src/shared/hooks/schedule/index.ts
@@ -1,3 +1,0 @@
-export { useScheduleList } from './useScheduleList';
-export { useScheduleDetail } from './useScheduleDetail';
-export { useScheduleContent } from './useScheduleContent';
