+++ apps/graduate/src/shared/hooks/useNoticeDetail.ts
